refactor: StakeData mod
