+++ crates/types/src/lib.rs
@@ -3,13 +3,13 @@
 mod operator;
 mod scope;
 mod session_id;
-mod stake_chain_data;
 mod stake_chain_id;
+mod stake_data;
 mod wots;
 
 pub use operator::OperatorPubKey;
 pub use scope::Scope;
 pub use session_id::SessionId;
-pub use stake_chain_data::StakeData;
 pub use stake_chain_id::StakeChainId;
+pub use stake_data::StakeData;
 pub use wots::Wots256PublicKey;
